@@ -12,35 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/cdk": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/compiler": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.2",
-    "@angular/platform-browser": "^14.1.2",
-    "@angular/platform-browser-dynamic": "^14.1.2",
-    "@angular/router": "^14.1.2",
-    "crypto-js": "~4.1.1",
-    "rxjs": "~6.6.0",
-    "tslib": "~2.3.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^14.2.8",
+    "@angular/cdk": "^14.2.6",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/core": "^14.2.8",
+    "@angular/forms": "^14.2.8",
+    "@angular/material": "^14.2.6",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/router": "^14.2.8",
+    "crypto-js": "~4.1.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.2",
-    "@angular/cli": "^14.1.2",
-    "@angular/compiler-cli": "^14.1.2",
-    "@types/crypto-js": "~4.0.2",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "~12.11.1",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@angular/cli": "^14.2.7",
+    "@angular/compiler-cli": "^14.2.8",
+    "@types/crypto-js": "~4.1.1",
+    "@types/jasmine": "~4.3.0",
+    "@types/node": "~18.11.8",
+    "jasmine-core": "~4.5.0",
+    "karma": "~6.4.1",
+    "karma-chrome-launcher": "~3.1.1",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
     "sonar-scanner": "~3.1.0",
-    "typescript": "~4.7.4"
+    "typescript": "~4.8.4"
   }
 }
